refactor(agents): dedupe dead helper code
