+++ src/agents/code-mode.ts
@@ -22,6 +22,7 @@ import {
   isCodeModeControlTool,
   markCodeModeControlTool,
 } from "./code-mode-control-tools.js";
+import { toCodeModeJsonSafe } from "./code-mode-json.js";
 import {
   createCodeModeApiVirtualFiles,
   createCodeModeNamespaceRuntime,
@@ -287,37 +288,8 @@ function reserveActiveRunSlot(): () => void {
   };
 }
 
-function toJsonSafe(value: unknown): unknown {
-  if (value === undefined) {
-    return null;
-  }
-  try {
-    const serialized = JSON.stringify(value);
-    return serialized === undefined ? null : (JSON.parse(serialized) as unknown);
-  } catch {
-    if (value instanceof Error) {
-      return { name: value.name, message: value.message };
-    }
-    if (value === null) {
-      return null;
-    }
-    switch (typeof value) {
-      case "string":
-      case "number":
-      case "boolean":
-        return value;
-      case "bigint":
-      case "symbol":
-      case "function":
-        return String(value);
-      default:
-        return Object.prototype.toString.call(value);
-    }
-  }
-}
-
 function jsonByteLength(value: unknown): number {
-  return Buffer.byteLength(JSON.stringify(toJsonSafe(value)) ?? "null", "utf8");
+  return Buffer.byteLength(JSON.stringify(toCodeModeJsonSafe(value)) ?? "null", "utf8");
 }
 
 class CodeModeLimitError extends ToolInputError {
@@ -600,7 +572,7 @@ async function runBridgeRequest(params: {
         break;
       }
     }
-    return { id: params.request.id, ok: true, value: toJsonSafe(value) };
+    return { id: params.request.id, ok: true, value: toCodeModeJsonSafe(value) };
   } catch (error) {
     return { id: params.request.id, ok: false, error: errorMessage(error) };
   }
